Move exceptions to specific folder
